Prepare some mock data in JavaScript
For now, we will use this mock data to build the skeleton of our app. We assign the variable **countries** an *array of objects*, each *object* being one country, one table row.

> This is the most common way of working with D3, as it can create its *selections* only from *arrays* and not *objects*. In JavaScript, *objects* can look like *arrays* with string-referenced *keys*, also known as *associative arrays* e.g. in PHP. They do **not** however work that way, as they are not (easily) *iterable* the way true *arrays* are.

Back to **coutries**. Each *object* here has two properties, **name** (a *string* with the country name) and **medals**, an *array* with sums of medals won in individual years (i.e. Canada won 11 medals in 1998 and 2002, then 22 medals in 2006 and finally 24 medals in 2010). You can find the legend for the years in the variable **years**.

> As a rule of thumb, when you have some repetitive data (set of countries or years) that you will display as table rows, bars in a bar chart or datapoints in a graph, you will want them to be in an **array**. If on the other hand you have **distinct properties** like name, medal count or e.g. abbreviated name of a country, you will want them to be in an **object** (you will not iterate over them, but they will be in individual cells of a table or influence the color of the bar / datapoint).