now we create *ghost rows* by selecting all **tr** (table row) elements and assigning them data (the **countries** array) with selection.data command
This will not actually create the rows in the document (HTML), but will instead create "ghost rows" only in JavaScript. There is information on which rows need to be **added**, **removed** or **modified**, or entering, exiting and updating **selection**. I will get to the concept of entering and updating selections in just a minute.

> Understanding the concept of Selections is key to understanding D3 (and, sadly, also the biggest hurdle).