fill in the medal count column
the beginning is very similar to the title column - again, we select third column and change its HTML based on bound data. But this time, we need to compute the sum of the medals, which is precisely what we do in the for loop. Essentially, we take each elelement of the medals array and add it to the sum variable. Then, we return the sum.