Lets get started with graphis. First, scales
Here we setup scales for x- and y- axis. We use d3.scale function, linear one. We use two functions, domain and range. Domain specifies input data, that is what range of values can be displayed (minimum to maximum, that is the count of years and maximum of medals). Range on the other hand denotes dimensions of the output, in our case, the SVG element (160x30px).