Now add a function that will help us draw the sparkline of medals won
we use d3.svg.line() helper. The .x and .y functions specify, how the coordinates should be computed. For x, we take the index of the data and convert it with the x-scale function defined earlier. The same is done with y-coordinate, we just take the data value and convert it with y-scale.