Draw the actual sparkline (path)
This hunk contains the following: first, we select the fourth column and in it, the svg element. To this element we assign the datum (singular of data) of this row's .medals array. Then we append a path element to the svg and set its definition (the d attribute) to be computed by our helper line function.